requirements.txt
firebase_admin
fastapi
pydantic
google
pytest
pytest-cov
pyrebase4
python-dotenv
